Time: 8 ms (60.16%) | Memory: 8.3 MB (48.06%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -2,12 +2,12 @@ class Solution {
 public:
   bool isPalindrome(int x) {
 long reverse=0;
-int temp = x;
+long temp = x;
 
     if (x < 0) return false;
     while(temp != 0)   
     {  
-    int remainder = temp % 10;  
+    long remainder = temp % 10;  
     reverse = reverse * 10 + remainder;  
     temp = temp/10;  
     }
